move fields around
