+++ netket/utils/struct/__init__.py
@@ -14,4 +14,6 @@
 
 # flake8: noqa: F401
 
-from .dataclass import dataclass, field, property_cached, Uninitialized
+from .fields import field, static_field, property_cached, Uninitialized
+
+from .dataclass import dataclass
